requirements.txt
python-dotenv==1.0.1
pandas==2.2.3
matplotlib==3.10.1
seaborn==0.13.2
influxdb-client==1.48.0
scikit-learn==1.6.1
statsmodels==0.14.4
hmmlearn==0.3.3
loguru==0.7.3